OPFGenerator.E2ELRReserveScaler — Type

E2ELRReserveScaler

Samples reserve requirements following the procedure below:

Sample a minimum reserve requirement MRR from a uniform distribution U(lb, ub) (mrr_dist).

Compute the upper bound of reserve requirements for each generator as rmax = α * (pmax - pmin).

Fix the lower bound of reserve requirement per generator to zero.

Fix the reserve cost of each generator to zero.

The parameter α is a scaling factor that determines each generator's maximum reserve capacity. It is the factor parameter times the ratio of the largest generator's capacity to the sum of all generators' dispatchable capacity.

OPFGenerator.Glocal — Type

Glocal{G,L}

A glocal distribution with global/local factors α::G and η::L.

This distribution represents a random variable of the form ϵ = α×η, where

α is a scalar random variable, with distribution d_α::G

η is a vector random variable, with distribution d_η::L

α and η are independent random variables

OPFGenerator.LoadScaler — Type

LoadScaler{D}

Scales loads with multiplicative noise sample from d::D.

Scaling retains power factors, i.e., each load's active and reactive demand is scaled by the same number.

OPFGenerator.ScaledLogNormal — Method

ScaledLogNormal(l, u, σs)

Generate a Glocal distribution ϵ = α×η where α~U[l,u] and ηᵢ ~ LogNormal(-σᵢ²/2, σᵢ).

OPFGenerator.ScaledUniform — Method

ScaledUniform(l, u, σs)

Generate a Glocal distribution ϵ = α×η where α ~ U[l,u] and ηᵢ ~ U[1-σᵢ, 1+σᵢ].

OPFGenerator._dedupe_and_sort_h5! — Method

_dedupe_and_sort_h5!(D)

De-duplicated and sort dataset D in increasing order of random seeds.

Equivalent to _dedupe_h5!(D); _sort_h5!(D), but more efficient.

OPFGenerator._dedupe_h5! — Method

_dedupe_h5!(D)

De-duplicate points in h5 dataset D, according to their random seed.

OPFGenerator._merge_h5 — Method

_merge_h5(V::Vector{Array{T,N})

Concatenate a collection of N-dimensional arrays along their last dimension.

This function is semantically equivalent to cat(V...; dims=ndims(first(V))), but uses a more efficient, splatting-free, implementation. All elements of V must have the same size in the first N-1 dimensions.

OPFGenerator._select_h5! — Method

_select_h5!(D, p)

Select data points in D as indicated by p.

D should be a dictionary in h5-compatible format, and p is either a vector of indices, or a logical vector of the same length as D["meta"]["seed"].

If p is a vector of indices, then all values of p should be integers between 1 and the number of elements in D

If p is a logical vector, then it should have the same length as D["meta"]["seed"]. Only datapoints i for which p[i] is true are selected.

OPFGenerator._sort_h5! — Method

_sort_h5!(D)

Sort dataset D in increasing order of random seeds.

The dictionary D should be in h5-compatible format. It is modified in-place.

The function expects D["meta"]["seed"] to exist and be a Vector{Int}. An error is thrown if such an entry is not found.

OPFGenerator.bridges — Method

bridges(data)

Identify whether each branch is a bridge.

The input data must be in basic format.

A branch is a bridge if removing it renders the network disconnected. Returns a dictionary res::Dict{String,Bool} such that res[br] is true if branch br is a bridge, and false otherwise.

OPFGenerator.build_opf — Method

build_acopf(data, optimizer)

Build an AC-OPF model.

This implementation is based on the AC-OPF formulation of Rosetta-OPF https://github.com/lanl-ansi/rosetta-opf/blob/38a951326df3156d79dcdc49c8010aa29905b05d/jump.jl

OPFGenerator.build_opf — Method

build_dcopf(data, optimizer)

Build a DC-OPF model.

OPFGenerator.build_opf — Method

build_soc_opf(data, optimizer)

Build an SOC-OPF model.

This implementation is based on the SOC-OPF formulation of PMAnnex.jl https://github.com/lanl-ansi/PMAnnex.jl/blob/f303f3c3c61e2d1a050ee7651fa6e8abc4055b55/src/model/opf.jl

OPFGenerator.extract_result — Method

_extract_acopf_solution(model, data)

Extract ACOPF solution from optimization model. The model must have been solved before.

OPFGenerator.extract_result — Method

_extract_dcopf_solution(model, data)

Extract DCOPF solution from optimization model. The model must have been solved before.

OPFGenerator.extract_result — Method

_extract_solution(model, data)

Extract SOC-OPF solution from optimization model. The model must have been solved before.

OPFGenerator.load_h5 — Function

load_h5

OPFGenerator.load_json — Method

load_json(filename::AbstractString)

Load JSON data from file filename.

OPFGenerator.save_h5 — Method

save_h5(filename, D)

Saves dictionary D to HDF5 file filename.

All keys in D must be of String type, and it must be HDF5-compatible. Additional restrictions are enforced on the values of D, see below.

Warning

Only the following types are supported:

String

(un)signed integers up to 64-bit precision

Float32 and Float64

Complex versions of the above numeric types

Dense Arrays of the the above scalar types

Numerical data of an unsupported type will be converted to Float64 when possible. An error will be thrown if the conversion is not possible.

OPFGenerator.save_json — Method

save_json(filename::AbstractString, data; indent)

Save data into JSON file filename. The following formats are supported:

uncompressed JSON .json

Gzip-compressed JSON .json.gz

Bzip2-compressed JSON .json.bz2

If the file extension does not match one of the above, an error is thrown.

OPFGenerator.tensorize — Method

tensorize(V)

Concatenate elements of V into a higher-dimensional tensor.

Similar to Base.stack, with one major difference: if V is a vector of scalars, the result is a 2D array M whose last dimension is length(V), and such that M[:, i] == V[i].

This function is only defined for Vector{T} and Vector{Array{T,N}} inputs, to avoid any unexpected behavior of Base.stack.

Random.rand! — Method

rand!(rng::AbstractRNG, s::AbstractOPFSampler, data::Dict)

Sample one new OPF instance and modify data in-place.

data must be a Dict in PowerModels format, representing the same network (i.e., same grid components with same indexing) as the one used to create s.